De-dupe getDiffContextSnippet().
